requirements.txt
flask
flask-mysqldb
Flask-SQLAlchemy
flask-login
Werkzeug==2.0.0
pyyaml